Updated Pickups and Spawning
Pickups now have a fault value set so when spawned it knows about it.  Created a Spawner that randomly creates a Pickup and places it randomly in the scene.  I did want them to fall from the sky but the mesh is all fucked and shit...  Need to compress it into single mesh or something.
Pickups will destroy themselves after 60 seconds.
HUD has some debig info about spawn count.